@@ -2,8 +2,17 @@ import {NgModule} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
 import {HomeComponent} from "./home/home.component";
 import {CaptureComponent} from "./capture/capture.component";
-import {UserResolve} from "./common/user/user.resolve";
+import {UserResolve} from "./shared/user/user.resolve";
 import {CaptureResolve} from "./capture/capture.resolve";
+import {LoginComponent} from "./login/login.component";
+import {AuthGuard} from "./shared/auth/auth.guard";
+import {LoginGuard} from "./login/login.guard";
+import {AccountComponent} from "./account/account.component";
+import {ShopComponent} from "./shop/shop.component";
+import {ShopItemResolve} from "./shop/item/shop-item.resolve";
+import {CountItemResolve} from "./shop/count-item.resolve";
+import {InventoryComponent} from "./inventory/inventory.component";
+import {ItemListResolve} from "./inventory/item-list.resolve";
 
 
 const routes: Routes = [
@@ -12,7 +21,39 @@ const routes: Routes = [
     component: HomeComponent,
     resolve: {
       user: UserResolve
+    },
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'login',
+    component: LoginComponent,
+    canActivate: [LoginGuard]
+  },
+  {
+    path: 'account',
+    component: AccountComponent,
+    resolve: {
+      user: UserResolve
     }
+  },
+  {
+    path: 'shop',
+    component: ShopComponent,
+    resolve: {
+      user: UserResolve,
+      shopItems: ShopItemResolve,
+      countItem: CountItemResolve
+    },
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'inventory',
+    component: InventoryComponent,
+    resolve: {
+      user: UserResolve,
+      items: ItemListResolve
+    },
+    canActivate: [AuthGuard]
   },
   {
     path: 'capture',
