@@ -3,17 +3,17 @@ import {ActivatedRouteSnapshot, Resolve} from "@angular/router";
 import {Observable} from "rxjs";
 import {Capture} from "./capture.interface";
 import {CaptureService} from "./capture.service";
-import {LoginService} from "../home/login/login.service";
+import {AuthService} from "../shared/auth/auth.service";
 
 @Injectable()
 export class CaptureResolve implements Resolve<Capture[]> {
 
   constructor(private captureService: CaptureService,
-              private loginService: LoginService) {
+              private authService: AuthService) {
   }
 
   resolve(route: ActivatedRouteSnapshot): Observable<Capture[]> {
-    if (this.loginService.checkCredentials())
+    if (this.authService.isLoggedIn())
       return this.captureService.listCaptures(11);
     else
       return null;
