@@ -1,28 +1,27 @@
-import {LoginService} from "./login/login.service";
 import {Component} from "@angular/core";
-import {User} from "../common/user/user.interface";
+import {User} from "../shared/user/user.interface";
 import {ActivatedRoute} from "@angular/router";
+import {AuthService} from "../shared/auth/auth.service";
+import {HeaderService} from "../shared/header/header.service";
+import {UserService} from "../shared/user/user.service";
 
 @Component({
   selector: 'home',
   templateUrl: './home.component.html',
+  styleUrls: ['./home.component.sass']
 })
 
 export class HomeComponent {
-  isLoggedIn = false;
   user: User;
 
-  constructor(private loginService: LoginService,
+  constructor(private headerService: HeaderService,
+    private userService: UserService,
+    private authService: AuthService,
               private route: ActivatedRoute) {
+    this.headerService.showHeader("Llama Breeder", true);
   }
 
   ngOnInit() {
     this.user = this.route.snapshot.data.user;
-    console.log(this.user);
-    this.isLoggedIn = this.loginService.checkCredentials();
-    let i = window.location.href.indexOf('access_token');
-    if (!this.isLoggedIn && i != -1) {
-      this.loginService.retrieveToken();
-    }
   }
 }
