@@ -1,4 +1,4 @@
-package hrpg.server.common.exception;
+package hrpg.server.user.service.exception;
 
 //todo move to user package
 public class InsufficientCoinsException extends Exception {
